+++ RELEASE-NOTES.md
@@ -28,6 +28,7 @@
 * #238  Exp.mapVal(..) / mapConditionVal(..) to allow mapping nulls
 * #239 Upgrade json-path lib to 2.9.0
 * #242 DataFrame.getColumn(int) must throw declared IllegalArgumentException
+* #247 RowSet.locate(..) / index(..)
 
 ## Release 0.18